scoreboard still bug

+++ Assets/Scripts/Scoreboard/Score.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Score : MonoBehaviour
 {
   public string name;
-  public float score;
+  public int score;
 
-  public Score(string name, float score)
+  public Score(string name, int score)
   {
     this.name = name;
     this.score = score;
@@ -18,7 +18,7 @@ public class Score : MonoBehaviour
     return this.name;
   }
 
-  public float getScore()
+  public int getScore()
   {
     return this.score;
   }
